Made kill run only once with stop per instance.

+++ ScreenToWebcam.sh
@@ -38,8 +38,10 @@ startS2W () {
 
 stopS2W () {
 	echo "Stopping ScreenToWebcam."
-	xargs kill -s SIGTERM < $ffmpegpidfile
-	tail --pid=$(cat $ffmpegpidfile) -f /dev/null
+	if (xargs ps < $ffmpegpidfile) > /dev/null ; then
+		xargs kill -s SIGTERM < $ffmpegpidfile
+		tail --pid=$(cat $ffmpegpidfile) -f /dev/null
+	fi
 	sudo modprobe -r v4l2loopback
 	modprobeExitCode=$?
 	if [ $modprobeExitCode -eq 0 ] ; then
